move super smart query to std/tests
